Avoid passing None to rclpy.init
Otherwise, all CLI arguments will be parsed as ROS arguments, which can lead
to rcl warnings or incorrect legacy remapping behavior.

This change does not pass arguments to rclpy.init from any of the CLI
tools, but it leaves the opportunity to do so in the future by setting the
parser argument 'argv'. For example, we could take the remaining arguments
and pass them to rclpy.init, similar to what is done in ros2run:

https://github.com/ros2/ros2cli/blob/4c5d9327026ecb2ea10a16b3429908b4f6f64ca6/ros2run/ros2run/command/run.py#L51-L53

Fixes #336.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>