[ros2doctor] Only report topic warnings if publisher or subscription count is zero
Before the tool was generating many false positives as it is a valid configuration
to have an unequal number of publishers and subscriptions on a given topic.

This change makes it so we only provide a warning if one of the counts is zero and the
other is not. Although this is still a legitimate configuration, it seems more likely to be
a problem and worth reporting.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>