Use reliable QoS for ros2topic tests
The TestROS2TopicCLI tests perform feature testing of the ros2topic
command line interface. If the system is under stress during these
tests, messages may be lost (by design). If that happens, there is a
fairly high likelihood that the test_topic_pub_once test will fail
because there is only one opportunity for the message to be successfully
transported. We're likely dropping other messages in this suite, but the
other tests continuously publish until one of the messages is received
(or a timeout occurs), making them significantly less likely to fail.

Using the 'reliable' setting for QoS reliability seems to make the tests
consistently pass, even when the system is placed under additional
stress.

Signed-off-by: Scott K Logan <logans@cottsay.net>