[Foxy backport] add --keep-alive option to 'topic pub' (#544)  use transient_local and longer keep-alive for pub tests (#546) Use reliable QoS for ros2topic tests (#555) (#565)
* add --keep-alive option to 'topic pub' (#544)

Signed-off-by: Dirk Thomas <dirk-thomas@users.noreply.github.com>
Signed-off-by: Shane Loretz <sloretz@osrfoundation.org>

* Give kwarg default value for backporting

keep_alive is made a keyword argument with a default value of 0.1 so the
pull request can be backported.

Signed-off-by: Shane Loretz<sloretz@openrobotics.org>
Signed-off-by: Shane Loretz <sloretz@osrfoundation.org>

* use transient_local and longer keep-alive for pub tests (#546)

* use transient_local and longer keep-alive for pub tests

Signed-off-by: Dirk Thomas <dirk-thomas@users.noreply.github.com>

* add comment to document unit of --keep-alive

Signed-off-by: Dirk Thomas <dirk-thomas@users.noreply.github.com>
Signed-off-by: Shane Loretz <sloretz@osrfoundation.org>

* Use reliable QoS for ros2topic tests (#555)

The TestROS2TopicCLI tests perform feature testing of the ros2topic
command line interface. If the system is under stress during these
tests, messages may be lost (by design). If that happens, there is a
fairly high likelihood that the test_topic_pub_once test will fail
because there is only one opportunity for the message to be successfully
transported. We're likely dropping other messages in this suite, but the
other tests continuously publish until one of the messages is received
(or a timeout occurs), making them significantly less likely to fail.

Using the 'reliable' setting for QoS reliability seems to make the tests
consistently pass, even when the system is placed under additional
stress.

Signed-off-by: Scott K Logan <logans@cottsay.net>
Signed-off-by: Shane Loretz <sloretz@osrfoundation.org>

Co-authored-by: Dirk Thomas <dirk-thomas@users.noreply.github.com>
Co-authored-by: Scott K Logan <logans@cottsay.net>