[ros2pkg] Skip copyright test in ament_cmake pkg
The template sources populated through `ros2 pkg create` for the
`ament_cmake` build type do not have a well-formed copyright header
and the package itself does not have a proper license file. As such, the
copyright and cpplint linters report missing license/copyright problems
with the generated package.

This commit skips said tests by default, instead informing the user to
comment out the lines which allow the copyright and cpplint linter
checks to be skipped once license/copyright headers are placed in the
package appropriately.

Signed-off-by: Abrar Rahman Protyasha <aprotyas@u.rochester.edu>