Fix how ros2 param interprets command-line arguments.
Basically ensure that we are *always* using the parsed yaml
string, so that users can override types as appropriate.
This makes it possible for users to specify something like:

ros2 param set /node param '!!str off'

for force a string to be "off" as opposed to having YAML
intepret it as a boolean.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>