Switch ros2 doctor to using psutil for network checks. (#687)
ifcfg has a couple of different problems.  The first is that
the code literally calls 'ifconfig' and parses the output.  While
this generally works, it is pretty ugly.  For those reasons,
Debian has specifically excluded it.

Instead, switch to psutil to get information about the network
interfaces on the system (at least, on POSIX).  This gets us
all of the information we had before, but using ioctl's instead.

Note that we can't actually get any flags from psutil at the
moment.  There is a PR open upstream to get this information,
but since it is not going to be available soon we implement
the equivalent thing in Python by using fcntl.ioctl.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>