Adds a `--max-wait-time` option to `ros2 topic pub`  (#800)
* Adds a `--max-wait-time` option to `ros2 topic pub`

When we run `ros2 topic pub -w 1 ...` we can wait for a fixed number
of connections prior to shutting down. This is useful in many scenarios,
however when writing automated tests it may be useful to time out if the
correct number of connections are not recieved in a given amount of
time. This PR adds a `--max-wait-time` option which allows users to set
a timeout on how long to wait for subscribers.

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>
Signed-off-by: Chris Lalancette <clalancette@gmail.com>