Fix the SIGTERM handling in the ros2 daemon. (#887)
* Fix the SIGTERM handling in the ros2 daemon.

There are 2 fixes needed here:
1.  Make sure to check rclpy.ok() in the overall loop
for handling requests.  That's because rclpy is handling
the signals by default.
2.  Realize that server.handle_request() uses the set
timeout as a timeout to select(), essentially.  Because of
that, we instead set the timeout to a short value (200 milliseconds),
and do the overall timeout checking ourselves.

Note that because we are waking up more often, this will cause
the ros2 daemon to use more CPU than before.  But it should
be negligible overall.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>
Co-authored-by: Scott K Logan <logans@cottsay.net>