Allow reuse address when firing up the ros2cli daemon.
Especially in tests where the daemon is being repeatedly
created and destroyed, it can take some time for the kernel
to actually allow the address to be rebinded (even after the
old process has exited).  This can lead to some situations
where we fail to spawn the daemon.

To fix this, set "allow_reuse_address" inside the LocalXMLRPCServer,
which will set SO_REUSADDR on the socket.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>