Fix stringifying InterfaceFlags when the flags are empty. (#1026)
It can be the case (especially on Windows) that the flags
for an interface are empty.  In that case, when stringifying,
we should return an empty string, not "nothing".  The latter
causes an exception and makes the NetworkReport fail to
work.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>
(cherry picked from commit 74f6989e71ed0376ff026d03a347acf2314d0c32)